refactor: redis repository 분리
