+++ src/main/java/com/ridetogether/server/domain/member/domain/Member.java
@@ -14,6 +14,7 @@ import com.ridetogether.server.domain.report.domain.Report;
 import com.ridetogether.server.global.common.BaseTimeEntity;
 import jakarta.persistence.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -86,7 +87,11 @@ public class Member extends BaseTimeEntity {
 
 	@OneToMany(mappedBy = "member")
 	@JsonIgnore
-	private List<MemberMatching> memberMatching;
+	private List<MemberMatching> memberMatching = new ArrayList<>();
+
+	public void addMemberMatching(MemberMatching memberMatching) {
+		this.memberMatching.add(memberMatching);
+	}
 
 	public void updateRefreshToken(String refreshToken) {
 		this.refreshToken = refreshToken;
