+++ src/main/java/com/ridetogether/server/global/config/AppConfig.java
@@ -1,0 +1,8 @@
+package com.ridetogether.server.global.config;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AppConfig {
+
+}
