+++ src/main/java/com/ridetogether/server/global/config/CorsProperties.java
@@ -1,0 +1,15 @@
+package com.ridetogether.server.global.config;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@Getter
+@Builder
+@ConfigurationProperties(prefix = "cors")
+public class CorsProperties {
+    private String allowedOrigins;
+    private String allowedMethods;
+    private String allowedHeaders;
+    private Long maxAge;
+}
