+++ src/main/java/com/ridetogether/server/global/security/domain/CustomUserDetails.java
@@ -6,6 +6,8 @@ import jakarta.persistence.FetchType;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.Map;
+
 import lombok.Builder;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.AuthorityUtils;
